break looker into smaller modules
