+++ app/packages/looker/src/lookers/image-looker.ts
@@ -1,0 +1,107 @@
+import { AbstractLooker } from "..";
+import { LookerUtils } from "./shared";
+import { getImageElements } from "../elements";
+import { COMMON_SHORTCUTS } from "../elements/common";
+import { Overlay } from "../overlays/base";
+import { DEFAULT_IMAGE_OPTIONS, ImageState, Optional } from "../state";
+
+import { zoomToContent } from "../zoom";
+
+export { createColorGenerator, getRGB } from "@fiftyone/utilities";
+export { freeVideos } from "../elements/util";
+export type {
+  Coloring,
+  FrameConfig,
+  FrameOptions,
+  ImageConfig,
+  ImageOptions,
+  Point,
+  VideoConfig,
+  VideoOptions,
+} from "../state";
+export { zoomAspectRatio } from "../zoom";
+
+export class ImageLooker extends AbstractLooker<ImageState> {
+  getElements(config) {
+    return getImageElements(config, this.updater, this.getDispatchEvent());
+  }
+
+  getInitialState(
+    config: ImageState["config"],
+    options: ImageState["options"]
+  ) {
+    options = {
+      ...this.getDefaultOptions(),
+      ...options,
+    };
+
+    return {
+      ...this.getInitialBaseState(),
+      config: { ...config },
+      options,
+      SHORTCUTS: COMMON_SHORTCUTS,
+    };
+  }
+
+  getDefaultOptions() {
+    return DEFAULT_IMAGE_OPTIONS;
+  }
+
+  hasDefaultZoom(state: ImageState, overlays: Overlay<ImageState>[]): boolean {
+    let pan = [0, 0];
+    let scale = 1;
+
+    if (state.options.zoom) {
+      const zoomState = zoomToContent(state, overlays);
+      pan = zoomState.pan;
+      scale = zoomState.scale;
+    }
+
+    return (
+      scale === state.scale &&
+      pan[0] === state.pan[0] &&
+      pan[1] === state.pan[1]
+    );
+  }
+
+  postProcess(): ImageState {
+    if (!this.state.setZoom) {
+      this.state.setZoom = this.hasResized();
+    }
+
+    if (this.state.zoomToContent) {
+      LookerUtils.toggleZoom(this.state, this.currentOverlays);
+    } else if (this.state.setZoom && this.state.overlaysPrepared) {
+      if (this.state.options.zoom) {
+        this.state = zoomToContent(this.state, this.pluckedOverlays);
+      } else {
+        this.state.pan = [0, 0];
+        this.state.scale = 1;
+      }
+
+      this.state.setZoom = false;
+    }
+
+    return super.postProcess();
+  }
+
+  updateOptions(options: Optional<ImageState["options"]>) {
+    const reload = LookerUtils.shouldReloadSample(this.state.options, options);
+    const state: Optional<ImageState> = { options };
+    if (options.zoom !== undefined) {
+      state.setZoom =
+        this.state.options.zoom !== options.zoom || this.state.config.thumbnail;
+    }
+
+    if (reload) {
+      this.updater({
+        ...state,
+        reloading: this.state.disabled,
+        disabled: false,
+      });
+      this.updateSample(this.sample);
+    } else {
+      this.updater({ ...state, disabled: false });
+    }
+  }
+}
